Relocated TrainingModel to training package
