+++ src/main/java/com/example/posganize/models/membership/MembershipModel.java
@@ -1,6 +1,6 @@
 package com.example.posganize.models.membership;
 
-import com.example.posganize.models.TrainingModel;
+import com.example.posganize.models.training.TrainingModel;
 import com.example.posganize.models.user.UserMembershipModel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
